refactor(SaleLotto) : LottoSeller을 Service로 수정
- 기존의 Lotto 도메인과 Payment 도메인을 가지고 흐름만 제어한다고 판단.
- 판매하는 행위를 Service로 관리, 좀 더 직관적인 이름으로 LottoPurchaseService로 변경
