+++ src/main/java/lotto/Application.java
@@ -1,12 +1,17 @@
 package lotto;
 
 import lotto.controller.LottoController;
+import lotto.domain.lottogenerator.RandomLottoGenerator;
+import lotto.service.LottoPurchaseService;
 import lotto.view.InputView;
 import lotto.view.OutputView;
 
 public class Application {
     public static void main(String[] args) {
-        LottoController lottoController = new LottoController(new InputView(), new OutputView());
+        LottoController lottoController = new LottoController(
+                new InputView(),
+                new OutputView(),
+                new LottoPurchaseService(new RandomLottoGenerator()));
         lottoController.run();
     }
 }
