+++ src/main/java/lotto/service/LottoPurchaseService.java
@@ -1,22 +1,24 @@
-package lotto.domain;
+package lotto.service;
 
 import lotto.config.LottoInfo;
+import lotto.domain.Lotto;
+import lotto.domain.Payment;
 import lotto.domain.lottogenerator.LottoGenerator;
 
 import java.util.List;
 import java.util.stream.Stream;
 
-public class LottoSeller {
+public class LottoPurchaseService {
     private final LottoGenerator lottoGenerator;
-    public LottoSeller(LottoGenerator lottoGenerator) {
+    public LottoPurchaseService(LottoGenerator lottoGenerator) {
         this.lottoGenerator = lottoGenerator;
     }
 
-    public List<Lotto> sellLottos(Payment payment) {
+    public List<Lotto> purchase(Payment payment){
         return Stream.generate(lottoGenerator::generate)
                 .limit(calculatePurchasableCount(payment))
                 .toList();
-    }
+    };
 
     private int calculatePurchasableCount(Payment payment) {
         return payment.getValue() / LottoInfo.LOTTO_PRICE;
